decimal

+++ DTOs/HiddenGems/CreateHiddenGemDto.cs
@@ -5,6 +5,6 @@ namespace TravelAPI.DTOs.HiddenGems
         public string Name { get; set; } = null!;
         public string Description { get; set; } = null!;
         public string Image { get; set; } = null!;
-        public double Rating { get; set; }
+        public decimal Rating { get; set; }
     }
 }
